[refactor/#11] api/supabse.ts에 있던 함수들 util로 분리
