+++ src/app/consult/[consultId]/page.tsx
@@ -1,6 +1,7 @@
 "use client";
 // 상담내역 상세페이지[3-2-1. 의사 답변이 달리기 전에 질문자 질문만 있는 세부페이지 ]
-import { getAnswerDetail, getConsultDetail, supabase } from "@/api/supabase";
+import { supabase } from "@/api/supabase";
+import { getConsultDetail, getAnswerDetail } from "@/hooks/getConsultData";
 import { useQuery } from "@tanstack/react-query";
 import { useEffect, useState } from "react";
 import { useRouter } from "next/navigation";
