+++ src/components/selftest/Departments.tsx
@@ -76,7 +76,6 @@ const Departments: React.FC<DepartmentsProps> = ({ onSelectDepartment }) => {
                 }
                 className="opacity-0 w-0 h-0"
               />
-
               <div>
                 {selectedDepartment === department ? (
                   <CheckedIcon />
