+++ tsconfig.json
@@ -8,7 +8,6 @@
     "noImplicitAny": false,
     "noEmit": true,
     "esModuleInterop": true,
-    "noImplicitAny": false,
     "module": "esnext",
     "moduleResolution": "bundler",
     "resolveJsonModule": true,
